feat(moved the page singleTrip to the folder Pages) WIP
