+++ src/app/app-routing.module.ts
@@ -7,7 +7,7 @@ import { LoginComponent } from './pages/auth/login/login.component';
 import { SignupComponent } from './pages/auth/signup/signup.component';
 import { NavbarComponent } from './shared/layout/navbar/navbar.component';
 import { JoinOurTeamComponent } from './pages/join-our-team/join-our-team.component';
-import { SingleTripComponent } from './trips/single-trip/single-trip.component';
+import { SingleTripComponent } from './pages/single-trip/single-trip.component';
 
 const routes: Routes = [
   { path: '', component: HomeContentComponent },
@@ -25,4 +25,4 @@ const routes: Routes = [
   imports: [RouterModule.forRoot(routes)],
   exports: [RouterModule],
 })
-export class AppRoutingModule { }
+export class AppRoutingModule {}
