+++ src/app/app.module.ts
@@ -17,7 +17,7 @@ import { LoginComponent } from './pages/auth/login/login.component';
 import { SignupComponent } from './pages/auth/signup/signup.component';
 import { JoinOurTeamComponent } from './pages/join-our-team/join-our-team.component';
 import { TripCardComponent } from './shared/cards/trip-card/trip-card.component';
-import { SingleTripComponent } from './trips/single-trip/single-trip.component';
+import { SingleTripComponent } from './pages/single-trip/single-trip.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
 @NgModule({
@@ -32,7 +32,7 @@ import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
     SignupComponent,
     JoinOurTeamComponent,
     TripCardComponent,
-    SingleTripComponent
+    SingleTripComponent,
   ],
   imports: [
     BrowserModule,
@@ -46,4 +46,4 @@ import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
   providers: [],
   bootstrap: [AppComponent],
 })
-export class AppModule { }
+export class AppModule {}
