+++ src/app/pages/single-trip/single-trip.component.scss
