+++ src/app/pages/single-trip/single-trip.component.ts
